Create new SessionPage component
